app/pipeline/service-crawling/crawling/l3-scan/jade.io-legislation/requirements.txt
requests
beautifulsoup4
PyYAML
mysql-connector-python
python-dotenv
boto3
selenium
webdriver-manager